+++ server_all.go
@@ -189,7 +189,7 @@ func (sc *Server) Read() (*Message, error) {
 // Write - writes a non multipart message to the ipc Connection.
 // msgType - denotes the type of data being sent. 0 is a reserved type for internal messages and errors.
 //
-func (connection *Connection) Write(msgType int, message []byte) error {
+func (connection *Connection) Write(msgType int32, message []byte) error {
 
 	if msgType == 0 {
 		return errors.New("Message type 0 is reserved")
@@ -230,7 +230,7 @@ func (sc *Server) write(connection *Connection) {
 
 		toSend = append(toSend, m.Data...)
 
-		writer.Write(intToBytes(len(toSend)))
+		writer.Write(intToBytes(int32(len(toSend))))
 		writer.Write(toSend)
 
 		err := writer.Flush()
